Stroages: Separate ColumnFileTinyReader & ColumnFileSetInputStream
